# services/market-data/src/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/market/status": {
            "get": {
                "description": "Returns the current trading phase and session times for a given market (US or HK).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quote"
                ],
                "summary": "Get market trading status",
                "parameters": [
                    {
                        "enum": [
                            "US",
                            "HK"
                        ],
                        "type": "string",
                        "default": "US",
                        "description": "Market identifier (US or HK)",
                        "name": "market",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.MarketStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/quote": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest quote snapshot for a given symbol. For guest users the price is delayed 15 minutes and labeled accordingly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quote"
                ],
                "summary": "Get real-time quote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g. AAPL, 00700)",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.Quote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Full-text search across stock symbols and company names. Supports US (NYSE/NASDAQ) and HK (HKEX) markets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search stocks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (symbol prefix or company name keyword)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "US",
                            "HK"
                        ],
                        "type": "string",
                        "description": "Filter by market",
                        "name": "market",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_search.Stock"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search/hot": {
            "get": {
                "description": "Returns the top-N most searched symbols in the last 24 hours, ranked by search frequency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Get hot search rankings",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of top symbols to return",
                        "name": "n",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_search.HotSearchItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.Market": {
            "type": "string",
            "enum": [
                "US",
                "HK"
            ],
            "x-enum-comments": {
                "MarketHK": "HKEX",
                "MarketUS": "NYSE/NASDAQ"
            },
            "x-enum-descriptions": [
                "NYSE/NASDAQ",
                "HKEX"
            ],
            "x-enum-varnames": [
                "MarketUS",
                "MarketHK"
            ]
        },
        "github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.MarketStatus": {
            "type": "object",
            "properties": {
                "market": {
                    "$ref": "#/definitions/github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.Market"
                },
                "phase": {
                    "$ref": "#/definitions/github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.TradingPhase"
                },
                "updatedAt": {
                    "description": "Always UTC",
                    "type": "string"
                }
            }
        },
        "github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.Quote": {
            "type": "object",
            "properties": {
                "ask": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "askSize": {
                    "type": "integer",
                    "format": "int64"
                },
                "bid": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "bidSize": {
                    "type": "integer",
                    "format": "int64"
                },
                "change": {
                    "description": "shopspring/decimal — never float64 — Price - PrevClose (unadjusted)",
                    "type": "number"
                },
                "changePct": {
                    "description": "shopspring/decimal — never float64 — Change / PrevClose * 100",
                    "type": "number"
                },
                "high": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "isStale": {
                    "description": "true if quote exceeds stale threshold",
                    "type": "boolean"
                },
                "lastUpdatedAt": {
                    "description": "Always UTC",
                    "type": "string"
                },
                "low": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "market": {
                    "$ref": "#/definitions/github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.Market"
                },
                "open": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "prevClose": {
                    "description": "shopspring/decimal — never float64 — previous Regular Session close",
                    "type": "number"
                },
                "price": {
                    "description": "shopspring/decimal — never float64",
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "github_com_hxiaodon_ai-broker_services_market-data_internal_quote_domain.TradingPhase": {
            "type": "string",
            "enum": [
                "PRE_MARKET",
                "OPEN",
                "LUNCH_BREAK",
                "POST_MARKET",
                "CLOSED"
            ],
            "x-enum-comments": {
                "PhaseLunchBreak": "HK only"
            },
            "x-enum-descriptions": [
                "",
                "",
                "HK only",
                "",
                ""
            ],
            "x-enum-varnames": [
                "PhasePreMarket",
                "PhaseOpen",
                "PhaseLunchBreak",
                "PhasePostMarket",
                "PhaseClosed"
            ]
        },
...
        "internal_search.HotSearchItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "score": {
                    "description": "Redis sorted set score — display ranking only, not financial",
                    "type": "number",
                    "format": "float64"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "internal_search.Stock": {
            "type": "object",
            "properties": {
                "exchange": {
                    "description": "\"NYSE\", \"NASDAQ\", \"HKEX\"",
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "market": {
                    "description": "\"US\" or \"HK\"",
                    "type": "string"
                },
                "name": {
                    "description": "Company name",
                    "type": "string"
                },
                "nameCN": {
                    "description": "Chinese company name",
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
    }
}